+++ vllm/assets/video.py
@@ -10,7 +10,7 @@ import numpy.typing as npt
 from huggingface_hub import hf_hub_download
 from PIL import Image
 
-from vllm.utils import PlaceholderModule
+from vllm.utils.importlib import PlaceholderModule
 
 from .base import get_cache_dir
