+++ vllm/attention/backends/registry.py
@@ -4,7 +4,7 @@
 
 import enum
 
-from vllm.utils import resolve_obj_by_qualname
+from vllm.utils.importlib import resolve_obj_by_qualname
 
 
 class _Backend(enum.Enum):
