+++ vllm/compilation/backends.py
@@ -24,7 +24,8 @@ from vllm.compilation.partition_rules import (
 from vllm.config import CompilationConfig, CUDAGraphMode, VllmConfig
 from vllm.logger import init_logger
 from vllm.platforms import current_platform
-from vllm.utils import is_torch_equal_or_newer, resolve_obj_by_qualname
+from vllm.utils import is_torch_equal_or_newer
+from vllm.utils.importlib import resolve_obj_by_qualname
 
 from .caching import VllmSerializableFunction
 from .compiler_interface import (
