+++ vllm/entrypoints/openai/speech_to_text.py
@@ -32,7 +32,7 @@ from vllm.inputs.data import PromptType
 from vllm.logger import init_logger
 from vllm.model_executor.models import SupportsTranscription
 from vllm.outputs import RequestOutput
-from vllm.utils import PlaceholderModule
+from vllm.utils.importlib import PlaceholderModule
 
 try:
     import librosa
