+++ vllm/entrypoints/openai/tool_parsers/abstract_tool_parser.py
@@ -12,8 +12,8 @@ from vllm.entrypoints.openai.protocol import (
 )
 from vllm.logger import init_logger
 from vllm.transformers_utils.tokenizer import AnyTokenizer
-from vllm.utils import import_from_path
 from vllm.utils.collections import is_list_of
+from vllm.utils.importlib import import_from_path
 
 logger = init_logger(__name__)
