+++ vllm/lora/punica_wrapper/punica_selector.py
@@ -3,7 +3,7 @@
 
 from vllm.logger import init_logger
 from vllm.platforms import current_platform
-from vllm.utils import resolve_obj_by_qualname
+from vllm.utils.importlib import resolve_obj_by_qualname
 
 from .punica_base import PunicaWrapperBase
