+++ vllm/model_executor/model_loader/weight_utils.py
@@ -34,7 +34,7 @@ from vllm.model_executor.layers.quantization import (
     get_quantization_config,
 )
 from vllm.platforms import current_platform
-from vllm.utils import PlaceholderModule
+from vllm.utils.importlib import PlaceholderModule
 
 try:
     from runai_model_streamer import SafetensorsStreamer
