+++ vllm/reasoning/abs_reasoning_parsers.py
@@ -8,8 +8,8 @@ from functools import cached_property
 from typing import TYPE_CHECKING, Any
 
 from vllm.logger import init_logger
-from vllm.utils import import_from_path
 from vllm.utils.collections import is_list_of
+from vllm.utils.importlib import import_from_path
 
 if TYPE_CHECKING:
     from vllm.entrypoints.openai.protocol import (
