+++ vllm/transformers_utils/s3_utils.py
@@ -4,7 +4,7 @@
 import fnmatch
 from typing import TYPE_CHECKING, Optional
 
-from vllm.utils import PlaceholderModule
+from vllm.utils.importlib import PlaceholderModule
 
 if TYPE_CHECKING:
     from botocore.client import BaseClient
