+++ vllm/v1/structured_output/backend_lm_format_enforcer.py
@@ -10,7 +10,7 @@ import torch
 from transformers import PreTrainedTokenizerBase
 
 from vllm.sampling_params import SamplingParams
-from vllm.utils import LazyLoader
+from vllm.utils.importlib import LazyLoader
 from vllm.v1.structured_output.backend_types import (
     StructuredOutputBackend,
     StructuredOutputGrammar,
