+++ vllm/v1/structured_output/utils.py
@@ -13,7 +13,7 @@ from diskcache import Cache
 
 import vllm.envs as envs
 from vllm.logger import init_logger
-from vllm.utils import LazyLoader
+from vllm.utils.importlib import LazyLoader
 
 if TYPE_CHECKING:
     import outlines_core as oc
